Remove link attribute; `cargo test` now works
Rust no longer requires #[link] to link to external libraries.
build.rs completely takes care of linking.

If you attempt to run `cargo test`, it will fail to compile.
The compiler would complain about duplicate definitions.
This was a result of linking the library twice:
Once with #[link], and again with build.rs.